Update LLDB support for python 3
Make minor changes for Python 3 (e.g. change print statements to functions).
Modify the code to account for the different relationship of
bytes and string objects in Python 3.
Add comments to the accessor functions.
Add some inline help to the dylan-bt and dylan-break-gf add-in commands.
Improve flexibility of file locations in the dylan-lldb helper script.