dfmc-llvm-back-end: Fix C-callable-function returns
Since the <raw-machine-word> subtypes easily end up intermixed, the
LLVM type of the return value of a C-callable function may not match
the function's generated return type. This change adds a cast where
necessary to ensure that the operand of a ret instruction matches the
generated function's defined return type.

* sources/dfmc/llvm-back-end/llvm-emit-computation.dylan
  (do-emit-return-temporary on <llvm-value>): Determine the function's
   return type and cast the return value to match it if needed.