c-ffi: Work around LLVM c-callable limitations
The c-ffi tests call a Dylan C-callable wrapper using the FFI. Though
the C and HARP back-ends can deal with C-function-definer and
c-callable-wrapper-definer declaring the same C symbol in the same
compilation record, for the LLVM back-end the extern declaration
conflicts with the wrapper function definition. Since this is not a
common use case, it seems best to work around this issue by moving the
wrapper definitions to a separate compilation record.

* sources/lib/c-ffi/test/tests.dylan
  ($c-string-value, $c-string-value-setter, wilma): Move from here.

* sources/lib/c-ffi/test/tests-aux.dylan: New source file containing
  C-callable-wrapper definitions moved from tests.dylan

* sources/lib/c-ffi/test/c-ffi-test.lid: Reference tests-aux.dylan