build: Support releases with included clang, gc, and libunwind
Many operating system distributions include versions of LLVM/Clang,
the Boehm-Demers-Weiser garbage collector, and libunwind that are not
the most ideal for use with the Open Dylan compiler and runtime. This
change adds a script that will package an Open Dylan release along
with these other software packages configured for use with it.

* build/unix/release-with-batteries.sh: New script for building an
  Open Dylan release and packaging it with specified versions of
  clang+llvm, libgc, and libunwind.

* Makefile.in (dist): Make directory creation and removal use
  consistent command styles.