run-time: Properly transfer register values to unwind context
The original libunwind implementation uses ucontext_t as its context
type, so that the argument passed into a SA_SIGINFO signal handler
function can be passed directly to unw_init_local. Not all
implementations of the libinfo API follow this convention, however, so
to support these we need to transfer register values explicitly.

* sources/lib/run-time/stack-walker.c
  (dylan_dump_callstack): For platforms that require it, explicitly
   transfer register values from a passed-in signal context into the
   libunwind cursor for use in the back-trace.