collections: Rename the table macro to tabling
Having a macro in a frequently used library with a relatively common name, in
this case "table", can result in confusing compiler warnings:

```
catalog.dylan:254.12-14: Serious warning - Unexpected token "::".
                 --
          (table :: <table>, key :: <string>, expected-type :: <type>)
                 --
```

Technically the compiler could assume that something in parameter position MUST
NOT be a macro call, and similar with `let table = ...`, but this provides a
quick workaround.

I'm open to other names. I considered `make-table` as well and chose somewhat
arbitrarily.