common-dylan-test-suite: move code to better homes and gardens
This reorganizes common-dylan-test-suite into files that more-or-less match the
functionality being tested, instead of organizing by the type of definition
(macro, function, etc). This makes it so that tests for similar functionality
are colocated. It should make it easier to know whether the feature your
currently concerned with has decent test coverage, and to run the tests related
to a given feature.