dfmc-llvm-back-end: Fix the apply-mep entry-point definition
In commit febad5031c an attempt was made to factor out common code in
the apply-mep entry point for shifting arguments out of the final
sequence argument to be passed to the called function. The refactoring
was incorrect, however, because it did not take methods without #rest
arguments into account. This change restores the original
argument-shifting logic for now.

* sources/dfmc/llvm-back-end/llvm-entry-points.dylan
  (entry-point apply-mep): Separate out (again) the cases of methods
   without optionals, since it was incorrect to construct and pass a
   shifted optionals vector to these <simple-method> IEPs.