Makefile.in: enable 3-stage bootstrap using packages
The bootstrap stage 1 registry and the main registry are moved to the workspace
directory. So this is a new requirement, that the bootstrap must be done
inside a workspace. We can easily check for and create a workspace.json file in
the parent directory if necessary.

dylan-tool is built from source first, then the registry is updated using
dylan-tool update, then the bootstrap is much the same as always.