dfmc-optimization: Fix the self-call to loop transform
When possible, the optimizer will turn calls within a method body to
itself into loops. This requires replacing references to the function
parameters with references to the loop-merge, indirectly through a
type check. Previously the replacement did not include the new
loop-merge itself, but the loop-merge-argument counts as part of the
loop body and therefore may need to be substituted as well.

* sources/dfmc/optimization/entry-points.dylan
  (upgrade-self-call): Since references to the method parameter in
   the loop-merge cannot be replaced blindly, handle the
   loop-merge-argument specially.