Run channels tests as part of `make check`
Also renamed to channels-test-suite to match standard naming scheme.

The changes made to the Windows Makefile are untested. Also I don't believe
these tests are run on windows because it assumes that channels-tests is an
application, but in fact it never calls run-test-application. I didn't try to
fix that since I wouldn't be able to verify it.