mps: Use MVFF pool, not MV.
The MV pool was deprecated in 1.111 and removed in favor of
using MVFF (first fit) in 1.118. The MVFF pool was in 1.108 which
we still use on Windows, so it is safe to use here. This is required when
updating to newer versions of MPS (1.118 and later).

Additionally, MVFF got some performance enhancements in 1.114.