Disentangle dylan-test-suite and common-dylan-test-suite
Issue #1237

(1) When testworks used to require adding all tests to a suite in order to run
them it made some sense to have one test suite library use another test suite
library; since a specific test suite was always passed to run-test-application
there was no confusion. In the new regime, where it's normal to call
run-test-application with no arguments in order to run all defined tests, this
no longer works since the tests from the _used_ test suite library would also
be executed.

(1.1) By not passing any arguments to `run-test-application` in
system-test-suite-app, this change results in 25 more tests being run because
they were never added to any suite that was part of `system-test-suite`.

(2) The new library, common-dylan-test-utilities, is used by
{common-dylan,dylan,io,system}-test-suite. Where to put this library and what
to call it were somewhat arbitrary. An argument could be made that it should be
in sources/testing somewhere but this seemed simpler.

(3) Unfortunately, because Testworks' interface specification suites define
tests, it isn't possible to share the dylan module's specification in order to
(for example) introspect all the instantiable collection classes. Instead I
hard-coded the list. I consider this acceptable because it's unlikely to
change.