Replace make-dylan-app with dylan-tool
`make-dylan-app` is completely obsoleted by `dylan new application ...` now, so
remove it and update documentation and build scripts to use the latter.

The changes to the Windows Makefile are untested but are straight-forward. The
only problem I can think of is if the `dylan-tool` submodules don't get checked
out recursively.