website: Add "website" repo as a submodule
I tried to figure out how to copy the website repo here, with the git history,
and wasn't able to make it work. The conflicts in files like .gitignore and
README.rst are manageable but I ran into conflicts with git submodules that,
even though I did work through them, I have no confidence in the result being
correct. So here we are.