Add build support for .proto files
The user can either include the generated files into their own library or
generate a separate library. The former case is a bit of a hack:

* it requires that they use separate files for their library definition and
  module definition, so that

* the generated module definition files (and main code) can be inserted into
  the project between the first file (assumed to be the user's library
  definition) and the second file (usually the user's module definition). This
  way the user modules can use the generated protobuf modules and code.

Because of limitations in the build system, tools like protobuf-tool, motley,
scepter, et al are only invoked on "user projects". When a project is opened
via the registry it becomes a "system project". Therefore, compiling
sources/lib/protobuf-tool/tests/pbtool-test-one.lid must be done by passing the
.lid file to dylan-compiler directly, and that's why this commit does not
include a registry file for pbtool-test-one.

When generating a complete library, the user is responsible for creating a
registry for the generated library. In this commit, pbtool-test-two-pb is one
such registry file.

Fixing the build system (or just understanding it better) is future work.