system: Replace resolve-locator:locators with resolve-file:file-system
This is fundamentally a file-system operation so it was a mistake to put it in the
locators module in the first place. The "file" terminology matches that used in most
other file-system exports.

I've switched the way arguments are passed to the internal `%resolve-file` function
here. The internal function only deals with strings, instead of locators. This avoids
some unnecessary conversions when the user passes a string.